1. Added item damage system; 2. Fixed issue when mob make 2 sounds when attacked; 3. Changed Money system from Double to Long; 4. Changed display funds, now it show the short form with 'K', 'M', etc. suffixes

+++ src/main/java/com/phazerous/phazerous/economy/interfaces/IEconomySubscriber.java
@@ -3,5 +3,6 @@ package com.phazerous.phazerous.economy.interfaces;
 import java.util.UUID;
 
 public interface IEconomySubscriber {
-    void update(UUID playerUUID, double balance);
+
+    void update(UUID playerUUID, long balance);
 }
